projects/webserver/pyproject.toml
[project]
name = "webserver"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Giulio De Matteis", email = "hello@giuliodematte.is" }
]
dependencies = [
    "django>=5.0.7",
    "django-ninja>=1.2.1",
    "django-celery-results>=2.5.1",
    "celery>=5.4.0",
    "redis>=5.0.7",
    "pyscf>=2.6.2",
    "rdkit>=2024.3.3",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/webserver"]
